Move `repo` integration tests into `tests` folder
