+++ crates/gitbutler-repo/tests/mod.rs
@@ -1,3 +1,4 @@
 mod create_wd_tree;
 mod credentials;
 mod merge_base_octopussy;
+mod rebase;
